Revival
Group project for the course User Interface Programming I

We provide the implementation of a pub ordering system which fulfills four requirements: internationalization, drag and drop, undo-redo and responsive web design. There are two languages for this system, English and Swedish. Before user logs in, they can scan the whole menu. After logging in, user could click on the blocks to put products into the bill list. The bill list will show up in the right side. Moreover, the layout will be changed according to the screen size.

The project includes five folders, one HTML file, one LICENSE file and README file. The index.html is the main page for this project. All works is working in this file. LICENSE is the premission for the HTML copyright. The "bundle" folder consists of all languages library. "css" folder stores the css files we used. "flag" folder contains all countries' flags. "image" is the folder to store the used images. "js" folder contains all the needed Javascript files. In "js" folder, main.js is the main Javascript file which contains all the needed functions. The functionality is written in the main.js. All JS files are well documented.

There are 4 folders and 1 html files in this project. "js" folder store all the javascript files for this website.

Contributors: Group 17: Larsson, Petter; Murali, Suraj; Wang, Han